requirements.txt
Playwright
python-dotenv